requirements.txt
beautifulsoup4==4.11.1
pyTelegramBotAPI==4.9.0
requests==2.28.1
python-whois
tweepy
